system_stats: fixed alive/on not respond

+++ userbot/modules/system_stats.py
@@ -131,7 +131,6 @@ async def pipcheck(pip):
 @register(outgoing=True, pattern=r"^\.(?:alive|on)\s?(.)?")
 async def amireallyalive(alive):
     """ For .alive command, check if the bot is running.  """
-    logo = ALIVE_LOGO
     output = ("`My Detail Ubot `\n"
              f"`xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx `\n"
              f"•  ⚙️ `Telethon    : v{version.__version__} `\n"
@@ -139,9 +138,12 @@ async def amireallyalive(alive):
              f"•  👤 `User        :`  {DEFAULTUSER} \n"
              f"•  🐛 `Running on  : {UPSTREAM_REPO_BRANCH} `\n"
              f"`xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx `\n")
-
-    await bot.send_file(alive.chat_id, logo, caption=output)
-    await alive.delete()
+    if ALIVE_LOGO:
+        logo = ALIVE_LOGO
+        await bot.send_file(alive.chat_id, logo, caption=output)
+        await alive.delete()
+    else:
+        await alive.edit(output)
 
 @register(outgoing=True, pattern="^.aliveu")
 async def amireallyaliveuser(username):
